+++ src/main/kotlin/online/bingzi/luck/perms/bridge/spring/boot/starter/entity/result/PermissionCheckResult.kt
@@ -1,5 +1,7 @@
 package online.bingzi.luck.perms.bridge.spring.boot.starter.entity.result
 
+import com.fasterxml.jackson.annotation.JsonCreator
+import com.fasterxml.jackson.annotation.JsonProperty
 import online.bingzi.luck.perms.bridge.spring.boot.starter.entity.Node
 
 /**
@@ -8,7 +10,7 @@ import online.bingzi.luck.perms.bridge.spring.boot.starter.entity.Node
  * @property result 检查结果("true", "false", "undefined")
  * @property node 匹配的权限节点(如果有)
  */
-data class PermissionCheckResult(
-    val result: String,
-    val node: Node? = null
+data class PermissionCheckResult @JsonCreator constructor(
+    @JsonProperty("result") val result: String,
+    @JsonProperty("node") val node: Node? = null
 )